Move `user_confirmed` function to `main` module.
